passbook/oauth_client/requirements.txt
requests_oauthlib>=0.4.2
oauthlib>=2.0.6
